refactor: move ServerInfo into server_info.rs
